refactor(web): move buildPackZip to shared app/lib for reuse in ExportPanel
Agent Packs' zip helper is needed by the agent-generator and skills-generator
export panels too, so relocate it out of the agent-packs-scoped lib folder.
